--- a/chronos.cabal
+++ b/chronos.cabal
@@ -90,6 +90,7 @@ benchmark bench
     , QuickCheck
     , thyme
     , time
+    , vector
   default-language:    Haskell2010
 
 source-repository head