I've always wondered about Zepto. Turns out it is fast:

```
Benchmark bench: RUNNING...
benchmarking parsing/Time.parseTimeM
time                 8.935 μs   (8.447 μs .. 9.549 μs)
                     0.983 R²   (0.972 R² .. 0.998 R²)
mean                 8.538 μs   (8.367 μs .. 8.900 μs)
std dev              764.0 ns   (464.9 ns .. 1.190 μs)
variance introduced by outliers: 84% (severely inflated)

benchmarking parsing/Thyme.parseTime
time                 1.770 μs   (1.678 μs .. 1.851 μs)
                     0.989 R²   (0.985 R² .. 0.996 R²)
mean                 1.675 μs   (1.641 μs .. 1.720 μs)
std dev              125.4 ns   (83.81 ns .. 171.0 ns)
variance introduced by outliers: 81% (severely inflated)

benchmarking parsing/Thyme.timeParser
time                 1.042 μs   (1.009 μs .. 1.088 μs)
                     0.990 R²   (0.985 R² .. 0.996 R²)
mean                 1.106 μs   (1.067 μs .. 1.164 μs)
std dev              166.2 ns   (108.8 ns .. 254.1 ns)
variance introduced by outliers: 95% (severely inflated)

benchmarking parsing/Chronos.parserUtf8_YmdHMS
time                 367.0 ns   (363.1 ns .. 371.9 ns)
                     0.998 R²   (0.995 R² .. 0.999 R²)
mean                 373.1 ns   (367.2 ns .. 384.6 ns)
std dev              26.70 ns   (16.93 ns .. 42.87 ns)
variance introduced by outliers: 82% (severely inflated)

benchmarking parsing/Chronos.zeptoUtf8_YmdHMS
time                 218.1 ns   (216.1 ns .. 221.7 ns)
                     0.998 R²   (0.996 R² .. 1.000 R²)
mean                 220.3 ns   (218.2 ns .. 224.3 ns)
std dev              9.417 ns   (4.888 ns .. 14.93 ns)
variance introduced by outliers: 62% (severely inflated)

Benchmark bench: FINISH
```