@andrewthad 2018-06-22
Nice! Could we use the name `zeptoUtf8_YmdHMS` instead of `zparserUtf8_YmdHMS`? Other than that, it looks good. I'm not sure what's gotten you into faster time parsing, but I'm glad there's someone else out there who cares.

Also, if you like fast parsing, you should look at https://github.com/andrewthad/packed/blob/master/src/Packed/Bytes/Parser.hs. I haven't totally stabilized on the API yet, and I keep running into problems with unboxed sums (they can cause crashes when the GC runs, but the GHC devs have been prompt in fixing these issues). But it's very fast and should be able to run generating little or no garbage (which CPSed parsers cannot really do).